bl_ocr/requirements.txt
docling>=2.5.0
requests>=2.31.0
Pillow>=10.0.0
tqdm>=4.66.0
gradio>=4.0.0
